refact(modularization/services): categorizing services as module
