+++ dashboard/src/modules/admins/components/dialogs/mutation/index.tsx
@@ -17,7 +17,7 @@ import {
     AdminCreateSchema,
     AdminType,
 } from "@marzneshin/modules/admins";
-import { ServicesField } from "@marzneshin/features/services";
+import { ServicesField } from "@marzneshin/modules/services";
 import { UsernameField } from "@marzneshin/modules/users";
 import {
     PasswordField,
