+++ dashboard/src/modules/services/components/dialogs/index.ts
