+++ dashboard/src/modules/services/components/tables/users/index.tsx
@@ -2,7 +2,7 @@
 import { FC } from "react";
 import { DoubleEntityTable } from "@marzneshin/features/entity-table";
 import { columns } from "./columns";
-import { fetchServiceUsers, type ServiceType } from "@marzneshin/features/services";
+import { fetchServiceUsers, type ServiceType } from "@marzneshin/modules/services";
 
 interface ServicesUsersTableProps {
     service: ServiceType
