+++ dashboard/src/modules/services/services/create-service.mutate.ts
@@ -1,4 +1,4 @@
-import { ServiceMutationType, ServicesQueryFetchKey } from "@marzneshin/features/services";
+import { ServiceMutationType, ServicesQueryFetchKey } from "@marzneshin/modules/services";
 import { useMutation } from "@tanstack/react-query";
 import { fetch, queryClient } from "@marzneshin/utils";
 import { toast } from "sonner";
