+++ dashboard/src/modules/users/components/tables/services/columns.tsx
@@ -4,7 +4,7 @@ import {
 } from "@marzneshin/features/entity-table"
 import i18n from "@marzneshin/features/i18n"
 import { Checkbox } from "@marzneshin/components"
-import { ServiceType } from "@marzneshin/features/services"
+import { ServiceType } from "@marzneshin/modules/services"
 
 export const columns: ColumnDef<ServiceType>[] = [
     {
