+++ dashboard/src/modules/users/components/tables/services/index.tsx
@@ -7,7 +7,7 @@ import {
     useUsersUpdateMutation,
 } from "@marzneshin/modules/users";
 import { useTranslation } from "react-i18next";
-import { fetchUserServices } from "@marzneshin/features/services";
+import { fetchUserServices } from "@marzneshin/modules/services";
 import { useState, useCallback, FC } from "react";
 
 interface UserServicesTableProps {
