+++ dashboard/src/routes/_dashboard/services/$serviceId/delete.tsx
@@ -5,7 +5,7 @@ import {
 import {
     ServicesDeleteConfirmationDialog,
     useRouterServiceContext,
-} from "@marzneshin/features/services";
+} from "@marzneshin/modules/services";
 import { useDialog } from "@marzneshin/hooks";
 
 const ServiceDelete = () => {
